src/image_watermark/image_watermark.py
"""
图像水印类 - 负责图像水印的嵌入和提取
"""

import torch
import yaml
from PIL import Image
from typing import Dict, Any, Optional, Union
import os


class ImageWatermark:
    """图像水印处理类"""
    
    def __init__(self, config_path: Optional[str] = None):
        """
        初始化图像水印处理器
        
        Args:
            config_path: 配置文件路径
        """
        self.config = self._load_config(config_path)
        self.algorithm = self.config.get('algorithm', 'stable_signature')
        self._setup_model()
    
    def _load_config(self, config_path: Optional[str]) -> Dict[str, Any]:
        """加载配置文件"""
        if config_path is None:
            # TODO: 使用默认配置
            return {
                'algorithm': 'stable_signature',
                'watermark_key': 'default_key',
                'model_name': 'runwayml/stable-diffusion-v1-5'
            }
        
        with open(config_path, 'r', encoding='utf-8') as f:
            config = yaml.safe_load(f)
        return config.get('image_watermark', {})
    
    def _setup_model(self):
        """初始化模型"""
        # TODO: Day 3, 6-7 实现具体的模型加载
        print(f"Setting up {self.algorithm} model...")
        self.model = None
        self.decoder = None
    
    def embed_watermark(self, image_input: Union[str, Image.Image], 
                       watermark_key: str = None, 
                       prompt: str = None) -> Image.Image:
        """
        在图像中嵌入水印
        
        Args:
            image_input: 输入图像（文件路径或PIL Image对象）
            watermark_key: 水印密钥
            prompt: 生成提示词（用于扩散模型）
            
        Returns:
            含水印的图像
        """
        # TODO: Day 6-7 实现具体的水印嵌入逻辑
        print(f"Embedding watermark in image using {self.algorithm}")
        
        if isinstance(image_input, str):
            image = Image.open(image_input)
        else:
            image = image_input
            
        # 占位符返回原图像
        return image
    
    def extract_watermark(self, image_input: Union[str, Image.Image], 
                         watermark_key: str = None) -> Dict[str, Any]:
        """
        从图像中提取水印
        
        Args:
            image_input: 含水印的图像（文件路径或PIL Image对象）
            watermark_key: 水印密钥
            
        Returns:
            提取结果，包含是否检测到水印等信息
        """
        # TODO: Day 6-7 实现具体的水印提取逻辑
        print(f"Extracting watermark from image using {self.algorithm}")
        return {
            'detected': False,
            'confidence': 0.0,
            'watermark_info': None
        }
    
    def generate_with_watermark(self, prompt: str, watermark_key: str = None) -> Image.Image:
        """
        生成带水印的图像
        
        Args:
            prompt: 生成提示词
            watermark_key: 水印密钥
            
        Returns:
            生成的含水印图像
        """
        # TODO: Day 6-7 实现具体的图像生成逻辑
        print(f"Generating image with watermark using {self.algorithm}")
        # 占位符返回空白图像
        return Image.new('RGB', (512, 512), color='white')
    
    def batch_embed(self, images: list, watermark_key: str = None) -> list:
        """批量嵌入水印"""
        return [self.embed_watermark(img, watermark_key) for img in images]
    
    def batch_extract(self, images: list, watermark_key: str = None) -> list:
        """批量提取水印"""
        return [self.extract_watermark(img, watermark_key) for img in images] 