requirements.txt
# 基础依赖
numpy>=1.21.0
torch>=1.10.0
torchvision>=0.11.0
transformers>=4.20.0
datasets>=2.0.0
Pillow>=9.0.0
matplotlib>=3.5.0
tqdm>=4.64.0

# 科学计算
scipy>=1.7.0
scikit-learn>=1.0.0

# 图像处理
opencv-python>=4.5.0
diffusers>=0.10.0

# 文本处理
nltk>=3.7
spacy>=3.4.0

# 工具
pyyaml>=6.0
requests>=2.25.0
psutil>=5.8.0

# 模型管理和加速
accelerate>=0.20.0
xformers>=0.0.20 